requirements_uv.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.2.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
bracex==2.4
build==1.2.1
certifi==2024.7.4
charset-normalizer==2.0.4
clarabel==0.9.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.2.1
cvxpy==1.5.2
cycler==0.12.1
diff-gaussian-rasterization @ git+https://github.com/dcharatan/diff-gaussian-rasterization-modified@1250c420ebb945f0dce9945086e22faab9157c92
dnspython==2.6.1
docker-pycreds==0.4.0
ecos==2.0.14
einops==0.8.0
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.53.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2024.6.1
gitdb==4.0.7
gitpython==3.1.43
gradio==4.37.2
gradio-client==1.0.2
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
importlib-metadata==7.0.1
importlib-resources==6.4.0
itsdangerous==2.2.0
jaxtyping==0.2.33
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kapture==1.1.10
kapture-localization==1.1.10
kiwisolver==1.4.5
lazy-loader==0.4
lightning==2.3.2
lightning-utilities==0.11.3.post0
llvmlite==0.43.0
loguru==0.5.3
lpips==0.1.4
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
networkx==3.3
numba==0.60.0
numpy==1.26.4
numpy-quaternion==2023.0.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.10.6
osqp==0.6.7.post0
packaging==24.1
pandas==2.2.2
piexif==1.1.3
pillow==10.3.0
pillow-heif==0.17.0
plotly==5.22.0
plyfile==1.0.3
poselib==2.0.2
protobuf==4.25.3
psutil==5.9.0
pycolmap==0.6.1
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
pyglet==1.5.29
pygments==2.18.0
pyopengl==3.1.0
pyparsing==3.2.5
pyproject-hooks==1.1.0
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
pywavelets==1.9.0
pyyaml==6.0.1
qdldl==0.1.7.post4
referencing==0.35.1
requests==2.32.2
retrying==1.3.3
rich==13.7.1
roma==1.5.0
rpds-py==0.19.0
ruff==0.5.1
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.5.1
scipy==1.14.0
scs==3.2.6
semantic-version==2.10.0
sentry-sdk==1.9.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==4.0.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.4.12
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1+cu121
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu121
tqdm==4.66.4
trimesh==4.4.3
triton==2.3.1
typeguard==2.13.3
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.16.6
watchfiles==0.22.0
wcmatch==8.5.2
websockets==11.0.3
werkzeug==3.0.3
yarl==1.9.3
zipp==3.17.0
